Splitted tests for types into different files
