+++ src/wheeled_model_enhanced/test/types/degree_test.cpp
@@ -1,0 +1,17 @@
+#include <gtest/gtest.h>
+
+#include "wheeled_model_enhanced/types/degree.hpp"
+
+TEST(types, degree)
+{
+    // Radian(double value)
+    {
+        constexpr double value = -5.231;
+        EXPECT_FLOAT_EQ(value, Degree(value).value());
+    }
+    // Radian(Degree value)
+    {
+        constexpr double value = 2.14675498;
+        EXPECT_FLOAT_EQ(123, Degree(Radian(value)).value());
+    }
+}
