+++ src/wheeled_model_enhanced/test/types_test.cpp
@@ -1,202 +1,6 @@
 #include <gtest/gtest.h>
 
 #include "wheeled_model_enhanced/types.hpp"
-#include "wheeled_model_enhanced/utils.hpp"
-
-TEST(types, radian)
-{
-    // Radian(double value)
-    {
-        constexpr double value = -5.231;
-        EXPECT_FLOAT_EQ(value, Radian(value).value());
-    }
-    // Radian(Degree value)
-    {
-        constexpr double value = 123;
-        EXPECT_FLOAT_EQ(2.14675498, Radian(Degree(value)).value());
-    }
-    // Radian normalize() const noexcept
-    {
-        constexpr double value = 7.783185307179586;
-        EXPECT_FLOAT_EQ(1.5, Radian(value).normalize().value());
-    }
-}
-
-TEST(types, degree)
-{
-    // Radian(double value)
-    {
-        constexpr double value = -5.231;
-        EXPECT_FLOAT_EQ(value, Degree(value).value());
-    }
-    // Radian(Degree value)
-    {
-        constexpr double value = 2.14675498;
-        EXPECT_FLOAT_EQ(123, Degree(Radian(value)).value());
-    }
-}
-
-TEST(types, meter)
-{
-    // Meter(double value)
-    {
-        constexpr double value = -5.231;
-        EXPECT_FLOAT_EQ(value, Meter(value).value());
-    }
-    // Meter(Kilometer value)
-    {
-        constexpr double value = -5.231;
-        EXPECT_FLOAT_EQ(-5231, Meter(Kilometer(value)).value());
-    }
-    // Meter &operator-(const Meter &rhv)
-    {
-        EXPECT_FLOAT_EQ(-0.25, (Meter(-0.125) - Meter(0.125)).value());
-    }
-    // Meter &operator-(double rhv)
-    {
-        EXPECT_FLOAT_EQ(-0.25, (Meter(-0.125) - 0.125).value());
-    }
-    // Meter &operator*(double rhv)
-    {
-        EXPECT_FLOAT_EQ(-0.25, (Meter(-0.125) * 2.f).value());
-    }
-    // Meter &operator-=(const Meter &rhv)
-    {
-        EXPECT_FLOAT_EQ(-0.25, (Meter(-0.125) -= 0.125).value());
-    }
-    // Meter &operator+(const Meter &rhv)
-    {
-        EXPECT_FLOAT_EQ(0, (Meter(-0.125) + Meter(0.125)).value());
-    }
-    // Meter &operator+(double rhv)
-    {
-        EXPECT_FLOAT_EQ(0, (Meter(-0.125) + 0.125).value());
-    }
-    // Meter &operator+=(const Meter &rhv)
-    {
-        EXPECT_FLOAT_EQ(0, (Meter(-0.125) += 0.125).value());
-    }
-    // double operator*(const double & lhv, const Meter & rhv)
-    {
-        EXPECT_FLOAT_EQ(-0.25, 2.f * Meter(-0.125));
-    }
-}
-
-TEST(types, kilometer)
-{
-    // Kilometer(double value)
-    {
-        constexpr double value = -5.231;
-        EXPECT_FLOAT_EQ(value, Kilometer(value).value());
-    }
-    // Kilometer(Meter value)
-    {
-        constexpr double value = -5.231;
-        EXPECT_FLOAT_EQ(-0.005231, Kilometer(Meter(value)).value());
-    }
-}
-
-TEST(types, pos)
-{
-    // constexpr Pos()
-    {
-        constexpr Pos pos;
-        EXPECT_FLOAT_EQ(0, pos.latitude().value());
-        EXPECT_FLOAT_EQ(0, pos.longitude().value());
-        EXPECT_FLOAT_EQ(0, pos.altitude().value());
-    }
-    // constexpr Pos(Radian latitude, Radian longitude, Meter altitude)
-    {
-        constexpr Pos pos(5.1, 7.2, 9.3);
-        EXPECT_FLOAT_EQ(5.1, pos.latitude().value());
-        EXPECT_FLOAT_EQ(7.2, pos.longitude().value());
-        EXPECT_FLOAT_EQ(9.3, pos.altitude().value());
-    }
-    // Pos(std::shared_ptr<const ReachGoalAction::Goal> goal)
-    {
-        using ReachGoalAction = wheeled_model_enhanced::action::ReachGoal;
-
-        const auto goal = std::make_shared<ReachGoalAction::Goal>();
-        goal->goal_lat = 123.22;
-        goal->goal_long = -23.2;
-
-        EXPECT_EQ(Pos(Degree(123.22), Degree(-23.2), Meter(0)), Pos(goal)) << Pos(goal);
-    }
-    // Pos &operator=(const sensor_msgs::msg::NavSatFix &msg)
-    {
-        Pos pos(1, 2, 3);
-        sensor_msgs::msg::NavSatFix msg;
-        msg.latitude = Degree(Radian(1.23)).value();
-        msg.longitude = Degree(Radian(2.11)).value();
-        msg.altitude = 122;
-
-        pos = msg;
-
-        EXPECT_FLOAT_EQ(1.23, pos.latitude().value());
-        EXPECT_FLOAT_EQ(2.11, pos.longitude().value());
-        EXPECT_FLOAT_EQ(122, pos.altitude().value());
-    }
-    // constexpr bool operator==(const Pos &rhv) const noexcept
-    {
-        constexpr Pos lhv(1, 2, 3);
-        constexpr Pos rhv(1, 2, 3);
-
-        EXPECT_TRUE(lhv == rhv);
-    }
-    {
-        constexpr Pos lhv(1, 2, 3);
-        constexpr Pos rhv(1, 2, 3.1);
-
-        EXPECT_FALSE(lhv == rhv);
-    }
-    // pos getters and setters
-    {
-        Pos pos(2.5, -1.2, 0.7);
-        EXPECT_FLOAT_EQ(2.5, pos.latitude().value()) << pos;
-        EXPECT_FLOAT_EQ(-1.2, pos.longitude().value()) << pos;
-        EXPECT_FLOAT_EQ(0.7, pos.altitude().value()) << pos;
-
-        pos.setLatitude(0.5);
-        pos.setLongitude(0.6);
-        pos.setAltitude(0.7);
-
-        EXPECT_FLOAT_EQ(0.5, pos.latitude().value());
-        EXPECT_FLOAT_EQ(0.6, pos.longitude().value());
-        EXPECT_FLOAT_EQ(0.7, pos.altitude().value());
-    }
-}
-
-TEST(types, cartesian)
-{
-    // Cartesian operator-(Cartesian lhv, const Cartesian &rhv)
-    {
-        Cartesian lhv(5, 7, 9);
-        Cartesian rhv(2, 3, 4);
-        EXPECT_EQ(Cartesian(3, 4, 5), lhv - rhv);
-        EXPECT_EQ(Cartesian(5, 7, 9), lhv) << lhv;
-    }
-    // Cartesian operator+(Cartesian lhv, const Cartesian &rhv)
-    {
-        Cartesian lhv(5, 7, 9);
-        Cartesian rhv(2, 3, 4);
-        EXPECT_EQ(Cartesian(7, 10, 13), lhv + rhv);
-        EXPECT_EQ(Cartesian(5, 7, 9), lhv) << lhv;
-    }
-}
-
-TEST(types, vector3d)
-{
-    // double operator*(const Vector3D &lhv, const Vector3D &rhv)
-    EXPECT_FLOAT_EQ(0, Vector3D(3, 1, -4) * Vector3D(8, -8, 4));
-
-    // Vector3D operator^(const Vector3D &lhv, const Vector3D &rhv)
-    EXPECT_EQ(Vector3D(-5, 10, 0), Vector3D(4, 2, -3) ^ Vector3D(2, 1, -4));
-}
-
-TEST(types, plane)
-{
-    EXPECT_EQ(Vector3D(0, 0, 300), Plane(0, 0, 300, 0).get_normal());
-}
 
 TEST(types, builtin_interfaces_msg_Time)
 {
@@ -228,136 +32,5 @@ TEST(types, builtin_interfaces_msg_Time)
 
         EXPECT_FLOAT_EQ(1.0087049007415771, lhv - rhv);
         EXPECT_FLOAT_EQ(-1.0087049007415771, rhv - lhv);
-    }
-}
-
-struct OptionalTest
-{
-    OptionalTest()
-    {
-    }
-    OptionalTest(bool ia, bool ib) : a(ia), b(ib)
-    {
-    }
-    bool a;
-    bool b;
-};
-
-struct OptionalDestructorTest
-{
-    OptionalDestructorTest(std::function<void()> callback) : _callback(callback)
-    {
-    }
-    ~OptionalDestructorTest()
-    {
-        _callback();
-    }
-    std::function<void()> _callback;
-};
-
-TEST(types, optional)
-{
-    // constructors
-    {
-        Optional<int> opt(3);
-
-        EXPECT_TRUE(opt.has_value());
-        EXPECT_EQ(3, *opt);
-    }
-    {
-        Optional<OptionalTest> opt({true, false});
-
-        EXPECT_TRUE(opt.has_value());
-        EXPECT_TRUE(opt->a);
-        EXPECT_FALSE(opt->b);
-    }
-    {
-        Optional<OptionalTest> opt;
-
-        EXPECT_FALSE(opt.has_value());
-        EXPECT_NO_THROW(opt->a);
-        EXPECT_NO_THROW(opt->b);
-    }
-    {
-        Optional<int> opt;
-
-        EXPECT_FALSE(opt.has_value());
-        EXPECT_NO_THROW(*opt);
-    }
-    {
-        const Optional<int> opt(5);
-        Optional<int> opt_copy(opt);
-
-        EXPECT_TRUE(opt.has_value());
-        EXPECT_EQ(5, *opt);
-        EXPECT_TRUE(opt_copy.has_value());
-        EXPECT_EQ(5, *opt_copy);
-    }
-    {
-        const Optional<int> opt;
-        Optional<int> opt_copy(opt);
-
-        EXPECT_FALSE(opt.has_value());
-        EXPECT_FALSE(opt_copy.has_value());
-    }
-    {
-        Optional<int> opt(5);
-        Optional<int> opt_copy(std::move(opt));
-
-        EXPECT_FALSE(opt.has_value());
-        EXPECT_TRUE(opt_copy.has_value());
-        EXPECT_EQ(5, *opt_copy);
-    }
-    // Optional<T> &operator=(const T &rhv)
-    {
-        Optional<int> opt;
-        int rhv = 3;
-        opt = rhv;
-
-        EXPECT_TRUE(opt.has_value());
-        EXPECT_EQ(3, *opt);
-    }
-    // constexpr Optional &operator=(const Optional &rhv)
-    {
-        Optional<int> opt(2);
-        Optional<int> rhv(7);
-
-        EXPECT_EQ(2, *opt);
-
-        opt = rhv;
-
-        EXPECT_EQ(7, *rhv);
-        EXPECT_EQ(7, *opt);
-    }
-    // void reset()
-    {
-        bool destructor_was_called = false;
-        const auto callback = [&destructor_was_called]() { destructor_was_called = true; };
-        Optional<OptionalDestructorTest> opt({callback});
-        opt.reset();
-
-        EXPECT_FALSE(opt.has_value());
-        EXPECT_TRUE(destructor_was_called);
-    }
-    // constexpr T &value()
-    {
-        constexpr Optional<int> opt;
-        EXPECT_THROW(opt.value(), BadOptionalAccess);
-    }
-    {
-        Optional<int> opt;
-        EXPECT_THROW(opt.value(), BadOptionalAccess);
-
-        opt = 7;
-        EXPECT_EQ(7, opt.value());
-    }
-    // constexpr const T &value() const
-    {
-        constexpr Optional<int> opt;
-        EXPECT_THROW(opt.value(), BadOptionalAccess);
-    }
-    {
-        constexpr Optional<int> opt(7);
-        EXPECT_EQ(7, opt.value());
     }
 }
